Extract GridLines
