+++ src/components/Canvas.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useRef } from 'react';
 import styled, { withTheme } from 'styled-components';
+import GridLines from './GridLines';
 
 const SquareAspectControl = styled.div`
   width: 100%;
@@ -23,63 +24,20 @@ const Styles = styled.div`
     width: 100%;
     ${'' /* border: 1px solid ${p => p.theme.colors.secondary}; */}
   }
-  canvas#canvas {
-    ${'' /* display: none; */}
-  }
 `;
-
-const dpr = window.devicePixelRatio || 1;
 
 function Canvas({ state, setCell, isPaused, theme }) {
   const canvasRef = useRef();
   const ctxRef = useRef();
-  const gridCanvasRef = useRef();
-  const gridCtxRef = useRef();
+  const dpr = window.devicePixelRatio || 1;
 
-  const setupGrid = () => {
-    // cells
+  const setup = () => {
     const c = canvasRef.current;
     const rect = c.getBoundingClientRect();
     c.width = rect.width * dpr;
     c.height = rect.width * dpr;
     ctxRef.current = c.getContext('2d');
     ctxRef.current.scale(dpr, dpr);
-
-    // grid
-    const g = gridCanvasRef.current;
-    const gridRect = g.getBoundingClientRect();
-    g.width = gridRect.width * dpr;
-    g.height = gridRect.width * dpr;
-    gridCtxRef.current = g.getContext('2d');
-    gridCtxRef.current.scale(dpr, dpr);
-
-    drawGridLines();
-  };
-
-  const drawGridLines = () => {
-    const c = gridCanvasRef.current;
-    const ctx = c.getContext('2d');
-    const cellSize = c.width / dpr / state.length;
-    const gridSize = c.width / dpr;
-
-    ctx.strokeStyle = theme.colors.secondary;
-    ctx.lineWidth = 1;
-
-    // perimeter
-    ctx.strokeRect(0, 0, gridSize, gridSize);
-    ctx.setLineDash([1, 1]);
-    for (let linePos = cellSize; linePos < gridSize - cellSize; linePos += cellSize) {
-      // rows
-      ctx.beginPath();
-      ctx.moveTo(0, linePos);
-      ctx.lineTo(gridSize, linePos);
-      ctx.stroke();
-      // cols
-      ctx.beginPath();
-      ctx.moveTo(linePos, 0);
-      ctx.lineTo(linePos, gridSize);
-      ctx.stroke();
-    }
   };
 
   const drawGrid = () => {
@@ -145,7 +103,7 @@ function Canvas({ state, setCell, isPaused, theme }) {
   };
 
   useEffect(() => {
-    setupGrid();
+    setup();
   }, []);
 
   useEffect(() => {
@@ -158,7 +116,7 @@ function Canvas({ state, setCell, isPaused, theme }) {
         <canvas id="canvas" ref={canvasRef} onMouseMove={handleMouseMove}>
           Your browser cannot display this content. :(
         </canvas>
-        <canvas id="grid-canvas" ref={gridCanvasRef} />
+        <GridLines dimension={state.length} />
       </Styles>
     </SquareAspectControl>
   );
